64-bit support, volume control, and general restructuring by [JAZ]

+++ bassmididrv_source/sound_out.h
@@ -6,7 +6,7 @@ class sound_out
 public:
 	virtual ~sound_out() {}
 
-	virtual const char* open( void * hwnd, unsigned sample_rate, unsigned nch, bool floating_point, unsigned max_samples_per_frame, unsigned num_frames ) = 0;
+	virtual const char* open( void * hwnd, unsigned sample_rate, unsigned short nch, bool floating_point, unsigned max_samples_per_frame, unsigned num_frames ) = 0;
 
 	virtual const char* write_frame( void * buffer, unsigned num_samples, bool wait = true ) = 0;
